+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Connection to local database, supporting Postgresql and sqlite
 - Command line parameters to configure connection to local database (-d for database enabled, -e for database engine, -r for database port, -u for database username, -w for database password)
+- Command line parameter to enable or disable local filesystem storage
 - Load command
 
 ### Changed